Create embl_1basic.r
We tried to explore the propertied of teh dataframe here. dimension of a dataframe, extraction of contents of the dataframe, names and vectors have been extracted from the dataframe. 

**Factors**
Factors represent the categorical data . It can contain a pre-defined set of values. By default R sort levels in alphabetical order.

***Manipulation of Dataframes***
There are a few functions for handling dataframes

**Pipe operator** : pass a dataframe to a function
**Filter operator** : Extract rows that fulfill certain criterias
** Select operator**: Select one or more columns of interest
**Grouped summaries**: Divide the dataframe into groups then the summarizing function will be applied on each group separately